# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: folicular
  description: |
    Backend service for Luteal: anonymous identity, offline-first delta sync,
    research-backed canonical data model, clearly labeled cycle estimates.

    This document is the contract's single source of truth. The Go server is
    validated against it (internal/contract tests); the Android client
    generates its DTOs from it. Prose documentation lives in docs/api.md.

    Conventions: dates are ISO-8601 calendar dates (format: date); instants
    are RFC 3339 UTC (format: date-time). Synchronized record IDs are
    client-generated UUIDv7 strings. Errors are RFC 9457 problem+json.
  version: 1.0.0
  contact:
    name: Luteal
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      operationId: register
      security: []
      summary: Create an anonymous account and register the first device
      description: |
        The account code is shown exactly once; only its SHA-256 hash is
        stored. Losing it means losing the account - no email recovery
        exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device_name: { type: string, maxLength: 120, example: Pixel 9 }
                invite_code:
                  type: string
                  description: >
                    Invite code required when registration is gated (closed
                    rollout). Ignored when registration is open.
                  example: LTL-BETA-0001
      responses:
        "201":
          description: Account and first device created
          content:
            application/json:
              schema:
                type: object
                required: [account, device, warning]
                properties:
                  account:
                    type: object
                    required: [id, code]
                    properties:
                      id: { $ref: "#/components/schemas/UUID" }
                      code:
                        type: string
                        description: Account code, shown once (LTL-XXXXX-XXXXX-XXXXX-XXXXX)
                        example: LTL-8K3FQ-Z2WNT-7HJMC-4XRDB
                  device: { $ref: "#/components/schemas/NewDevice" }
                  warning:
                    type: string
                    description: French user-facing notice about keeping the code safe
        default: { $ref: "#/components/responses/Problem" }
  /v1/auth/devices:
    post:
      tags: [auth]
      operationId: addDevice
      security: []
      summary: Register an additional device using the account code
      description: Rate limited per client IP. Failures are generic (no enumeration).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, example: LTL-8K3FQ-Z2WNT-7HJMC-4XRDB }
                device_name: { type: string, maxLength: 120 }
      responses:
        "201":
          description: Device registered
          content:
            application/json:
              schema:
                type: object
                required: [device]
                properties:
                  device: { $ref: "#/components/schemas/NewDevice" }
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    get:
      tags: [auth]
      operationId: listDevices
      summary: List the account's devices
      responses:
        "200":
          description: Device list
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items: { $ref: "#/components/schemas/Device" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/auth/devices/{deviceID}:
    delete:
      tags: [auth]
      operationId: revokeDevice
      summary: Revoke a device token
      parameters:
        - $ref: "#/components/parameters/DeviceID"
      responses:
        "204": { description: Device revoked }
        "404": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/me:
    get:
      tags: [account]
      operationId: getMe
      summary: Account, current device, and server-authoritative settings
      responses:
        "200":
          description: Account view
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        default: { $ref: "#/components/responses/Problem" }
    patch:
      tags: [account]
      operationId: patchMe
      summary: Partially update settings (server-authoritative, not synced)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsPatch" }
      responses:
        "200":
          description: Full account view after update
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "422": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/sync/push:
    post:
      tags: [sync]
      operationId: syncPush
      summary: Push a batch of client changes
      description: |
        Each change is validated and applied with an entity-level
        last-write-wins guard (updated_at, then client_rev). Changes that
        lose the guard are returned in `conflicts` carrying the server's
        current state - nothing is silently lost. Invalid changes are
        reported in `rejected`; valid changes in the same batch still apply.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [changes]
              properties:
                changes:
                  type: array
                  minItems: 1
                  maxItems: 1000
                  items: { $ref: "#/components/schemas/SyncChangeInput" }
      responses:
        "200":
          description: Batch results and new cursor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncPushResult" }
        "422": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/sync/pull:
    get:
      tags: [sync]
      operationId: syncPull
      summary: Pull ordered changes since a cursor
      parameters:
        - name: since
          in: query
          schema: { type: integer, format: int64, minimum: 0, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, format: int64, minimum: 1, maximum: 2000, default: 500 }
      responses:
        "200":
          description: Ordered changes (tombstones included)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncPullResult" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/cycles:
    get:
      tags: [reads]
      operationId: listCycles
      summary: Live cycles overlapping a date range
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Cycles, newest first
          content:
            application/json:
              schema:
                type: object
                required: [cycles]
                properties:
                  cycles:
                    type: array
                    items: { $ref: "#/components/schemas/CycleView" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/days:
    get:
      tags: [reads]
      operationId: listDays
      summary: Merged daily observation view for a date range
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Days with any observation, ascending
          content:
            application/json:
              schema:
                type: object
                required: [days]
                properties:
                  days:
                    type: array
                    items: { $ref: "#/components/schemas/DayView" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/predictions/current:
    get:
      tags: [estimates]
      operationId: predictionsCurrent
      summary: On-demand cycle estimates (ranges with uncertainty)
      description: |
        Computed from the account's own recorded cycle starts. With fewer
        than 3 starts, windows are null and no population default is
        substituted. Ovulation and fertile windows are capped at `low`
        confidence: calendar timing alone is a weak signal.
      responses:
        "200":
          description: Current estimates
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Prediction" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/duo/invitations:
    post:
      tags: [duo]
      operationId: createInvitation
      summary: Create a pending Duo link with a pairing code and URL
      description: |
        The pairing code (50 bits, 7-day expiry, single use) is transported
        typed, as pairing_url, or as a QR code the client renders from
        pairing_url. The code in the URL is a bearer-like secret.
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invitation" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/duo/links:
    post:
      tags: [duo]
      operationId: acceptLink
      security:
        - bearerAuth: []
      summary: Accept a pending invitation (partner's own account)
      description: Rate limited. Generic 404 prevents code enumeration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pairing_code]
              properties:
                pairing_code: { type: string, example: FQ7K2-WNT4H }
      responses:
        "201":
          description: Link activated
          content:
            application/json:
              schema:
                type: object
                required: [link]
                properties:
                  link: { $ref: "#/components/schemas/DuoLink" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    get:
      tags: [duo]
      operationId: listLinks
      summary: List the caller's Duo links in both roles
      responses:
        "200":
          description: Link list
          content:
            application/json:
              schema:
                type: object
                required: [links]
                properties:
                  links:
                    type: array
                    items: { $ref: "#/components/schemas/DuoLink" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/duo/links/{linkID}/grants:
    patch:
      tags: [duo]
      operationId: patchGrants
      summary: Grant or revoke one sharing field (tracker only)
      parameters:
        - $ref: "#/components/parameters/LinkID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [field, granted]
              properties:
                field: { $ref: "#/components/schemas/GrantField" }
                granted: { type: boolean }
      responses:
        "204": { description: Grant updated }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/duo/links/{linkID}:
    delete:
      tags: [duo]
      operationId: revokeLink
      summary: End a link (either member; also cancels pending invitations)
      parameters:
        - $ref: "#/components/parameters/LinkID"
      responses:
        "204": { description: Link revoked }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/duo/view:
    get:
      tags: [duo]
      operationId: duoView
      summary: Role-aware partner projection
      description: |
        Partner: only fields with an active grant; ungranted fields are
        null (absence deliberately indistinguishable from no data).
        Tracker: data fields null, support thread always visible. Private
        notes and raw observations never pass through this endpoint.
      responses:
        "200":
          description: Projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DuoView" }
        "404": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/duo/support-requests:
    post:
      tags: [duo]
      operationId: createSupportRequest
      summary: Ask for support (either member)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [link_id, kind]
              properties:
                link_id: { $ref: "#/components/schemas/UUID" }
                kind: { $ref: "#/components/schemas/SupportKind" }
                message: { type: string, maxLength: 1000 }
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupportRequest" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/duo/support-requests/{requestID}/ack:
    patch:
      tags: [duo]
      operationId: ackSupportRequest
      summary: Acknowledge a support request (recipient only)
      parameters:
        - name: requestID
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      responses:
        "204": { description: Acknowledged }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 256-bit device token issued at registration (ltok_...)
  parameters:
    DeviceID:
      name: deviceID
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
    LinkID:
      name: linkID
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
    From:
      name: from
      in: query
      description: Inclusive start date (defaults vary per endpoint)
      schema: { type: string, format: date }
    To:
      name: to
      in: query
      description: Inclusive end date
      schema: { type: string, format: date }
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    UUID:
      type: string
      format: uuid
      example: 019832e0-6c14-7000-8000-000000000001
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, example: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
    NewDevice:
      type: object
      required: [id, name, token]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        name: { type: string }
        token:
          type: string
          description: Bearer token, shown once (ltok_...)
    Device:
      type: object
      required: [id, name, created_at, revoked, current]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        name: { type: string }
        created_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time, nullable: true }
        revoked: { type: boolean }
        current: { type: boolean }
    LifeStage:
      type: string
      description: STRAW+10-aligned, user-selected, never inferred
      enum:
        - unknown
        - reproductive_early
        - reproductive_peak
        - reproductive_late
        - menopause_transition_early
        - menopause_transition_late
        - postmenopause_early
        - postmenopause_late
    TrackingFocus:
      type: string
      description: User-selected charting focus, never a diagnosis
      enum: [pms, pmdd, endometriosis, pcos, custom]
    Settings:
      type: object
      required: [locale, time_zone, life_stage, tracking_focus, updated_at]
      properties:
        locale: { type: string, example: fr }
        time_zone: { type: string, example: Europe/Paris }
        life_stage: { $ref: "#/components/schemas/LifeStage" }
        tracking_focus:
          type: array
          items: { $ref: "#/components/schemas/TrackingFocus" }
        updated_at: { type: string, format: date-time }
    SettingsPatch:
      type: object
      description: Partial settings update; merged state is validated as a whole
      properties:
        locale: { type: string }
        time_zone: { type: string }
        life_stage: { $ref: "#/components/schemas/LifeStage" }
        tracking_focus:
          type: array
          items: { $ref: "#/components/schemas/TrackingFocus" }
    Me:
      type: object
      required: [account, device, settings]
      properties:
        account:
          type: object
          required: [id, status, created_at]
          properties:
            id: { $ref: "#/components/schemas/UUID" }
            status: { type: string, enum: [active, disabled, deleted] }
            created_at: { type: string, format: date-time }
        device:
          type: object
          required: [id, name]
          properties:
            id: { $ref: "#/components/schemas/UUID" }
            name: { type: string }
        settings: { $ref: "#/components/schemas/Settings" }
    EntityType:
      type: string
      enum:
        - cycle
        - bleeding_observation
        - daily_entry
        - symptom_definition
        - symptom_log
        - biomarker_observation
        - medication_log
    SyncChangeInput:
      type: object
      required: [entity_type, data]
      properties:
        entity_type: { $ref: "#/components/schemas/EntityType" }
        data:
          type: object
          description: |
            The full record for entity_type, including the envelope (id,
            client_rev, created_at, updated_at, deleted_at). Shape per type:
            CycleData, BleedingObservationData, DailyEntryData,
            SymptomDefinitionData, SymptomLogData, BiomarkerObservationData,
            MedicationLogData. Deletes carry deleted_at set plus the last
            known data (tombstone semantics).
    AppliedChange:
      type: object
      required: [entity_type, entity_id, seq]
      properties:
        entity_type: { $ref: "#/components/schemas/EntityType" }
        entity_id: { $ref: "#/components/schemas/UUID" }
        seq: { type: integer, format: int64 }
    RejectedChange:
      type: object
      required: [entity_type, detail]
      properties:
        entity_type: { $ref: "#/components/schemas/EntityType" }
        entity_id: { $ref: "#/components/schemas/UUID" }
        detail: { type: string }
    ConflictChange:
      type: object
      required: [entity_type, entity_id, reason, current]
      properties:
        entity_type: { $ref: "#/components/schemas/EntityType" }
        entity_id: { $ref: "#/components/schemas/UUID" }
        reason: { type: string, enum: [superseded] }
        current:
          type: object
          description: The server's current record (adopt it; nothing is silently lost)
    SyncPushResult:
      type: object
      required: [applied, rejected, conflicts, cursor]
      properties:
        applied:
          type: array
          items: { $ref: "#/components/schemas/AppliedChange" }
        rejected:
          type: array
          items: { $ref: "#/components/schemas/RejectedChange" }
        conflicts:
          type: array
          items: { $ref: "#/components/schemas/ConflictChange" }
        cursor: { type: integer, format: int64 }
    SyncPullChange:
      type: object
      required: [seq, entity_type, entity_id, deleted, updated_at]
      properties:
        seq: { type: integer, format: int64 }
        entity_type: { $ref: "#/components/schemas/EntityType" }
        entity_id: { $ref: "#/components/schemas/UUID" }
        deleted: { type: boolean }
        updated_at: { type: string, format: date-time }
        data:
          type: object
          nullable: true
          description: Full record snapshot; null when deleted
    SyncPullResult:
      type: object
      required: [changes, cursor, has_more]
      properties:
        changes:
          type: array
          items: { $ref: "#/components/schemas/SyncPullChange" }
        cursor: { type: integer, format: int64 }
        has_more: { type: boolean }
    Flow:
      type: string
      description: Self-rated bleeding flow (FIGO-aligned vocabulary)
      enum: [none, spotting, light, medium, heavy]
    Certainty:
      type: string
      enum: [recorded, uncertain, estimated]
    RecordSource:
      type: string
      enum: [manual, import, estimated]
...
    CervicalFluid:
      type: string
      enum: [none, sticky, creamy, watery, egg_white, unresolved]
    CervixPosition:
      type: string
      enum: [low, medium, high, unknown]
    CervixFirmness:
      type: string
      enum: [firm, soft, unknown]
...
    MedicationKind:
      type: string
      enum: [contraceptive_hormonal, emergency_contraception, pain_relief, supplement, other]
...
    CycleView:
      type: object
      required: [id, start_date, certainty, source, notes, updated_at]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date, nullable: true }
        length_days: { type: integer, nullable: true }
        bleeding_days: { type: integer, nullable: true }
        certainty: { $ref: "#/components/schemas/Certainty" }
        source: { $ref: "#/components/schemas/RecordSource" }
        notes: { type: string }
        updated_at: { type: string, format: date-time }
    DayView:
      type: object
      required: [date, symptoms, medications]
      properties:
        date: { type: string, format: date }
        cycle_day: { type: integer, nullable: true }
        bleeding:
          type: object
          nullable: true
          required: [flow, intermenstrual]
          properties:
            flow: { $ref: "#/components/schemas/Flow" }
            intermenstrual: { type: boolean }
            product_count: { type: integer, nullable: true }
        entry:
          type: object
          nullable: true
          required: [notes]
          properties:
            pain_level: { type: integer, nullable: true }
            mood_level: { type: integer, nullable: true }
            energy_level: { type: integer, nullable: true }
            notes: { type: string }
        symptoms:
          type: array
          items:
            type: object
            required: [key, severity, notes]
            properties:
              key: { type: string }
              severity: { type: integer }
              notes: { type: string }
        biomarkers:
          type: object
          nullable: true
          properties:
            bbt_celsius: { type: number, format: double, nullable: true }
            cervical_fluid: { $ref: "#/components/schemas/CervicalFluid" }
            cervix_position: { $ref: "#/components/schemas/CervixPosition" }
            cervix_firmness: { $ref: "#/components/schemas/CervixFirmness" }
        medications:
          type: array
          items:
            type: object
            required: [name, dose, kind]
            properties:
              name: { type: string }
              dose: { type: string }
              kind: { $ref: "#/components/schemas/MedicationKind" }
    Confidence:
      type: string
      description: Uncertainty label; "certain" does not exist by design
      enum: [insufficient, low, moderate]
    Prediction:
      type: object
      required: [generated_at, method, basis, disclaimer]
      properties:
        generated_at: { type: string, format: date-time }
        method: { type: string, example: cycle_length_median_v1 }
        basis:
          type: object
          required: [cycle_count]
          properties:
            cycle_count: { type: integer }
            median_length_days: { type: integer }
            variability_days: { type: integer }
        next_menstruation:
          type: object
          nullable: true
          required: [window_start, central_date, window_end, confidence]
          properties:
            window_start: { type: string, format: date }
            central_date: { type: string, format: date }
            window_end: { type: string, format: date }
            confidence: { $ref: "#/components/schemas/Confidence" }
        ovulation_estimate:
          type: object
          nullable: true
          required: [window_start, central_date, window_end, confidence]
          properties:
            window_start: { type: string, format: date }
            central_date: { type: string, format: date }
            window_end: { type: string, format: date }
            confidence: { $ref: "#/components/schemas/Confidence" }
        fertile_window_estimate:
          type: object
          nullable: true
          required: [window_start, window_end, confidence]
          properties:
            window_start: { type: string, format: date }
            window_end: { type: string, format: date }
            confidence: { $ref: "#/components/schemas/Confidence" }
        disclaimer:
          type: string
          description: French server-authored non-medical disclaimer
    Invitation:
      type: object
      required: [link_id, pairing_code, pairing_url, expires_at]
      properties:
        link_id: { $ref: "#/components/schemas/UUID" }
        pairing_code: { type: string, example: FQ7K2-WNT4H }
        pairing_url:
          type: string
          description: Shareable link / QR payload / deep-link target
          example: https://luteal-duo.waldemar.site/accept?code=FQ7K2-WNT4H
        expires_at: { type: string, format: date-time }
    DuoRole:
      type: string
      enum: [tracker, partner]
    DuoLink:
      type: object
      required: [id, role, status, created_at]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        role: { $ref: "#/components/schemas/DuoRole" }
        status: { type: string, enum: [pending, active, revoked] }
        created_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time, nullable: true }
    GrantField:
      type: string
      enum: [cycle_day, period_estimate, mood, energy, support_requests]
    SupportKind:
      type: string
      enum: [general, comfort, practical, space]
    SupportRequest:
      type: object
      required: [id, author_role, kind, message, created_at]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        author_role: { $ref: "#/components/schemas/DuoRole" }
        kind: { $ref: "#/components/schemas/SupportKind" }
        message: { type: string }
        created_at: { type: string, format: date-time }
        acknowledged_at: { type: string, format: date-time, nullable: true }
    DuoView:
      type: object
      required: [link_id, role, as_of]
      properties:
        link_id: { $ref: "#/components/schemas/UUID" }
        role: { $ref: "#/components/schemas/DuoRole" }
        as_of: { type: string, format: date-time }
        cycle_day: { type: integer, nullable: true }
        period_estimate:
          type: object
          nullable: true
          required: [window_start, window_end]
          properties:
            window_start: { type: string, format: date }
            window_end: { type: string, format: date }
        mood: { $ref: "#/components/schemas/SharedLevel" }
        energy: { $ref: "#/components/schemas/SharedLevel" }
        support_requests:
          type: array
          nullable: true
          items: { $ref: "#/components/schemas/SupportRequest" }
    SharedLevel:
      type: object
      nullable: true
      required: [date, level]
      properties:
        date: { type: string, format: date }
        level: { type: integer, minimum: 1, maximum: 5 }